Changed type of LSN in tree-index pages from int to long.
git-svn-id: https://hyracks.googlecode.com/svn/branches/hyracks_btree_updates_next@652 123451ca-8445-de46-9d55-352943316053

+++ hyracks-storage-am-btree/src/main/java/edu/uci/ics/hyracks/storage/am/btree/impls/BTree.java
@@ -52,7 +52,7 @@ public class BTree implements ITreeIndex {
 
     public static final float DEFAULT_FILL_FACTOR = 0.7f;
 
-    private final static int RESTART_OP = Integer.MIN_VALUE;
+    private final static long RESTART_OP = Long.MIN_VALUE;
     private final static int MAX_RESTARTS = 10;
     private final static int rootPage = 1;
